Validations: Fix problems with h5py generated files
It seems the that the h5py doesn't create scalar string values. For the
most applications it seems not to be a problem, but the validation did
not handle this case, which leads to problems during the validation.

In case the expected value is a collection type (list, tuple,
numpy.ndarray) the first entry will taken as the string.